--- src/Lewee.Infrastructure.AspNet/Auth/AuthenticatedUserServiceConfiguration.cs
@@ -1,23 +1,0 @@
-﻿using Lewee.Domain;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Lewee.Infrastructure.AspNet.Auth;
-
-/// <summary>
-/// <see cref="AuthenticatedUserService"/> Configuration
-/// </summary>
-public static class AuthenticatedUserServiceConfiguration
-{
-    /// <summary>
-    /// Configures the <see cref="AuthenticatedUserService"/>
-    /// </summary>
-    /// <param name="services">Services collection</param>
-    /// <returns>Services collection for chaining</returns>
-    public static IServiceCollection AddAuthenticatedUserService(this IServiceCollection services)
-    {
-        services.AddHttpContextAccessor();
-        services.AddSingleton<IAuthenticatedUserService, AuthenticatedUserService>();
-
-        return services;
-    }
-}
