+++ src/Lewee.Infrastructure.Auth/AuthenticatedUserServiceConfiguration.cs
@@ -1,0 +1,23 @@
+﻿using Lewee.Domain;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Lewee.Infrastructure.Auth;
+
+/// <summary>
+/// <see cref="AuthenticatedUserService"/> Configuration
+/// </summary>
+public static class AuthenticatedUserServiceConfiguration
+{
+    /// <summary>
+    /// Configures the <see cref="AuthenticatedUserService"/>
+    /// </summary>
+    /// <param name="services">Services collection</param>
+    /// <returns>Services collection for chaining</returns>
+    public static IServiceCollection AddAuthenticatedUserService(this IServiceCollection services)
+    {
+        services.AddHttpContextAccessor();
+        services.AddSingleton<IAuthenticatedUserService, AuthenticatedUserService>();
+
+        return services;
+    }
+}
